Implemented basic permissions for commands
Permissions can be assigned per command in two ways. If a command is not defined globally or per user in the permissions file it will be allowed. If the command is allowed globally it will over ride the per user permissions. If the command is denied globally but allowed per user, it is allwoed. Users are represented by their user ID.