Revised permission infrastructure (#40)
* Revised permission infastructure

This revises and expands on the permissions object. Permissions are based on Discord server roles. Commands that are restricted must be associated with a guilds server role(s). Additional data members were added to the command base class which should be relevant to more than just the new commands this commit introduces. The message listener was also changed to allow the bot to respond to Direct Messages. A potentially breaking change was introduced by changing a variable name, but I believe the result is much clearer long term.

* Murdered comment

* Slightly less cruft

Relies on lazy evaluation and witchcraft to save on explicit if statements

* Simplified logic

Simplified the logic to determine whether the command runners roles allow them to run the command

* Convinced that it is acceptable to return early

* Added file extensions

* For loop budget cut, layoffs ensued

* Major refactoring of logic

Reduces cruft by not trying to return in just one location. By using some inverse logic we can get the fail cases out of the way first and focus on the main goal of the functions.

* Makes better use of the Collection methods

Makes better use of the Collection methods and clarifies the expected behavior.

* Checks the exception code instead of message

Try Catch blocks are places I really miss Types.

* Uses the new Logger object

* Uses new config file object

* Added usage method to command parent class

* Removed code this permisisons feature made useless

* Whitespace

* Made logger send more verbose

* Add logger abilty to message a user directly

* Allows bot to respond to all DM types

* Converted to Logger messaging; cleaned up requires

Converted to use the Logger object for all messaging. Because this required requiring the Logger object, I also fixed file extensions and pathing in the current command set.

* Cruft--

Uses ternary to get rid of an explicit if statement

* Murdered more comments

* Allow the permissions to be bypassed

Introduces a configuration file option to disable the permissions check on restricted commands. When set to true (default), permissions are checked on restricted commands to determine if they can be issued by the user. When set to false, all commands are able to be issued by anyone.

* Fixed more paths/file exts

* Whitespace

* Updated README for permissions